Update environment variables ready for deployment
This commit updates and separates the environment configurations for the different stages of the application workflow. A new environment file, `dev.env`, has been created to provide a set of standardized values for development.

In addition, the `nwserver.env` is renamed to `test.env` to make it clear that these settings are specific to the testing environment. New environment variables have been added in `test.env` to tune the software for testing scenarios.

Furthermore, `docker-compose.dev.yml` and `docker-compose.prod.yml` are updated to use the new environment files. A `prod.env` file has been created to cater to the production stage specifically.

These changes provide better separation of concerns between the development, testing, and production stages of the server, ensuring that each has its settings that are easy to manage and update as needed.